+++ app/Console/cake.php
@@ -26,7 +26,9 @@ if (function_exists('ini_set')) {
 
 	// the following line differs from its sibling
 	// /lib/Cake/Console/Templates/skel/Console/cake.php
-	ini_set('include_path', $root . $ds . 'lib' . PATH_SEPARATOR . ini_get('include_path'));
+	ini_set('include_path', $root . $ds . 'lib' .
+		PATH_SEPARATOR . $root . $ds . 'app/Lib'.
+		PATH_SEPARATOR . ini_get('include_path'));
 }
 
 if (!include($dispatcher)) {