+++ app/Controller/UsersController.php
@@ -1,5 +1,4 @@
 <?php
-include_once('RkLib.php');
 
 class UsersController extends AppController {
 	private $RK_access_token;